1) Fields : "Número, Complemento, Bairro" : Should be Obligatory fields
2) On "cadastra-se page" and on "finalizar page" : On "Dados profissionais" > Option " Acadêmico" > The field "Instituição de Ensino" is still double - just keep the first and delete the second

3) On "Finalizar compra" page, after i login, CPF is not appearing on the field even though i gave this info once already at first time of signup.
After i write the CPF and also choose the "DDD" code and i go back to the online shop to choose another product, when i came back to the "finalizar compra" page the infos of CPF and DDD given before are erased.
Also if i type in these infos and signout, when i sign in again these values are erased again. why is not saving as it did before?