1) On a new registration -on "cadastre-se" - If i try to finish cadastro without giving the "obligatory" infos message comes in german - to translate to : "Por favor, preencha os campos obligatorios"
Also, in this case, it shows only some obligatory fields which werent field up in red - i think if we show some in red than better all which are missing to be field up.

CPF filed was also not highlighted in red and its obligatory. Makes it confusing


4) On the back end : Vendas>Clientes>clientes>edit one client (for example: patomoon@web.de)
On tab " endereco" - there is only the street name saved - info i placed on cadstro on "endereco" field. But there is not the "número" and "complemento".
There is not even these fields there. Also funny that there is not showing the CPF